requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.3.35
langchain-mcp-adapters>=0.1.11
langgraph>=0.6.10
requests>=2.31.0
httpx>=0.25.0
pydantic>=2.0.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
cryptography>=3.4.8
openai>=1.0.0
tiktoken>=0.5.0
